+++ java/org/rocksdb/RocksDB.java
@@ -121,7 +121,6 @@ public class RocksDB extends RocksObject {
             options.numShardBits_, path);
     
     db.storeOptionsInstance(options);
-    
     return db;
   }